separate details body so it can be rendered in review as well
