+++ tutor/src/screens/assignment-edit/ux.js
@@ -332,13 +332,17 @@ export default class AssignmentUX {
     await this.saveAndCopyPlan();
   }
 
-  @action.bound async onPublishClick() {
+  @action.bound async savePlan(){
     this.plan.update(omit(this.form.values, 'tasking_plans'));
     this.plan.is_draft = false;
     if (!this.plan.is_published) {
       this.plan.is_publish_requested = true;
     }
     await this.saveAndCopyPlan();
+  }
+
+  @action.bound async onPublishClick() {
+    await this.savePlan();
     this.onComplete();
   }
